WEB - Usuarios Cadastro de Usuarios

+++ Entity/Pessoa.cs
@@ -8,8 +8,8 @@ namespace SistemaLoja01.Entity
     public class Pessoa
     {
         public string nome { get; set; }
-        public int cpf { get; set; }
-        public int contato { get; set; }
+        public long cpf { get; set; }
+        public long contato { get; set; }
         public string email { get; set; }
     }
 }
